GIF library: Declare image dimensions as unsigneds.
Not uint16_ts. Why? Because uint16_t promotes to int. So imagine
an image with "size" 49020x63995. This should produce an image
data array of size 3137034900. Which it does with unsigned numbers
for the dimensions. But uint16_t promotes to int, and 32-bitwise,
that size is negative.

+++ include/lcdfgif/gif.h
@@ -43,8 +43,8 @@ struct Gif_Stream {
     Gif_Colormap *global;
     uint8_t background;
 
-    uint16_t screen_width;
-    uint16_t screen_height;
+    unsigned screen_width;
+    unsigned screen_height;
     long loopcount;		/* -1 means no loop count */
 
     Gif_Comment *comment;
@@ -96,8 +96,8 @@ struct Gif_Image {
     uint16_t left;
     uint16_t top;
 
-    uint16_t width;
-    uint16_t height;
+    unsigned width;
+    unsigned height;
 
     uint8_t interlace;
     uint8_t **img;		/* img[y][x] == image byte (x,y) */
